rename header and remove unused files
